requirements.txt
langchain
wikipedia
python-dotenv
requests
pydantic
beautifulsoup4
langchain-anthropic
langchain-community
langchain-openai